requirements.txt
anthropic>=0.69.0
python-dotenv>=1.1.1